Move aes edek code to separate module
